pyproject.toml
[tool.poetry]
name = "sales-bot"
version = "0.1.0"
description = ""
authors = ["Krushovice <Kickstar69@yandex.ru>"]
readme = "README.md"
packages = [{include = "sales_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
aiosqlite = "^0.19.0"
flask = {extras = ["async"], version = "^3.0.2"}
requests = "^2.31.0"
outline-vpn-api = "^6.1.0"
yookassa = "^3.0.1"
alembic = "^1.13.1"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
